pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-gateway"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "httpx",
    "PyYAML",
    "pydantic",
    "uvicorn",
    "python-dotenv",
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "surrealdb>=1.0.7",
    "python-multipart>=0.0.21",
    "watchdog",
    "mcp",
    "tenacity>=9.1.2",
    "duckduckgo-search>=8.1.1",
    "mirascope>=1.0.0",
    "logfire[fastapi,httpx]>=4.17.0",
]

[tool.logfire]
# Suppress configuration warnings in development
ignore_no_config = true

[tool.setuptools.packages.find]
where = ["."]
include = ["agent_runner*", "common*"]
namespaces = true
